Refactor statistics
Rename folder and files to snake_case. Rename other fields, includes and
methods according to google style guide. Add EXPECT_TRUE, EXPECT_FALSE in
tests. Remove param from tests. Add const to some methods. Change types of
data in Statistics. Wrap tests for multithreads execution in #if #endif.
Now ColumnStats is a struct. Change order of access modifiers in CsvStats.
Delete move constructor and operator= from CsvStats. Change description
for some methods. Change Count to NumberOfValues, also change Count prefix
in other methods. Refactor ColumnStats::ToKeyValueStringPairs and it's
return value. Add first phase name for CsvStats. Change CsvStats::ShowSample
and CsvStats::DeleteNullAndEmpties. Add MakeFrom static method for DoubleType.
Change unsigned to size_t in CsvStats and TypedColumnData in some places.
Delete duplication of the code in Max/Min, Distinct/Quantile. Add virtual
method Clone to Type and it's successors.

+++ src/model/typed_column_data.cpp
@@ -73,9 +73,9 @@ TypedColumnData TypedColumnDataFactory::CreateMixedFromTypeMap(std::unique_ptr<T
 
     std::unordered_set<unsigned> const& nulls = type_map[TypeId::kNull];
     std::unordered_set<unsigned> const& empties = type_map[TypeId::kEmpty];
-    unsigned int const rows_num = unparsed_.size();
-    unsigned int const nulls_num = nulls.size();
-    unsigned int const empties_num = empties.size();
+    size_t const rows_num = unparsed_.size();
+    size_t const nulls_num = nulls.size();
+    size_t const empties_num = empties.size();
 
     TypeIdToType type_id_to_type = MapTypeIdsToTypes(type_map);
     size_t const buf_size = CalculateMixedBufSize(mixed_type, type_id_to_type, type_map);
@@ -112,9 +112,9 @@ TypedColumnData TypedColumnDataFactory::CreateConcreteFromTypeMap(std::unique_pt
 
     std::unordered_set<unsigned> nulls = std::move(type_map[TypeId::kNull]);
     std::unordered_set<unsigned> empties = std::move(type_map[TypeId::kEmpty]);
-    unsigned int const rows_num = unparsed_.size();
-    unsigned int const nulls_num = nulls.size();
-    unsigned int const empties_num = empties.size();
+    size_t const rows_num = unparsed_.size();
+    size_t const nulls_num = nulls.size();
+    size_t const empties_num = empties.size();
     assert(rows_num >= nulls_num + empties_num);
 
     std::vector<std::byte const*> data(unparsed_.size());
